requirements.txt
django
psycopg2
python-dotenv
drf-spectacular
django-ratelimit
djangorestframework
djangorestframework-simplejwt
